Split actions enum to the separate file
